requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.9.0
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
cryptography==41.0.7
aiofiles==23.2.1
scikit-learn==1.3.2
opencv-python
firebase-admin>=6.0.0
